@@ -31,32 +31,36 @@
   "homepage": "https://github.com/DHTMLX/optimus-start#readme",
   "devDependencies": {
     "babel-core": "^6.24.0",
-    "babel-loader": "^6.4.0",
+    "babel-loader": "^7.1.4",
     "babel-preset-es2015": "^6.24.0",
     "babel-preset-es2015-loose": "^8.0.0",
-    "css-loader": "^0.27.3",
-    "dhx-optimus": "^0.3.0",
-    "eslint": "^3.17.1",
+    "css-loader": "^0.28.11",
+    "dhx-optimus": "^0.4.0",
+    "eslint": "^4.19.1",
     "event-stream": "^3.3.4",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.10.1",
-    "gulp": "^3.9.1",
-    "gulp-clean": "^0.3.2",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^1.1.11",
+    "gulp": "^4.0.0",
+    "gulp-clean": "^0.4.0",
     "gulp-cssnano": "^2.1.2",
-    "gulp-eslint": "^3.0.1",
+    "gulp-eslint": "^4.0.2",
     "gulp-filter": "^5.0.0",
     "gulp-rename": "^1.2.2",
-    "gulp-replace": "^0.5.4",
-    "gulp-uglify": "^2.1.0",
+    "gulp-replace": "^1.0.0",
+    "gulp-uglify": "^3.0.0",
     "gulp-util": "^3.0.8",
     "handlebars": "^4.0.6",
     "handlebars-loader": "^1.4.0",
-    "less": "^2.7.2",
-    "less-loader": "^3.0.0",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.0.0",
-    "webpack-livereload-plugin": "^0.10.0",
-    "webpack-stream": "^3.2.0"
+    "less": "^3.0.4",
+    "less-loader": "^4.1.0",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-dev-server": "^3.1.4",
+    "webpack-livereload-plugin": "^2.1.1",
+    "webpack-stream": "^4.0.3"
+  },
+  "dependencies": {
+    "riot-route": "^3.1.3"
   }
 }
