@@ -1,33 +1,59 @@
-import {DHXView} from "dhx-optimus";
+import { DHXView } from 'dhx-optimus';
 
-export class SidebarView extends DHXView{
-	render(){
-		this.ui = this.root.attachSidebar({
-			template: "tiles",
-			icons_path: "codebase/imgs/"
-		});
-		this.ui.attachEvent("onSelect", (id) => this.callEvent("SideBar", [id]) );
-		this._load();
-	}
+export class SidebarView extends DHXView {
+  render() {
+    dhtmlXSideBar.prototype.templates.faTiles =
+      '<i><div class="dhxsidebar_item_icon #icon#"></div><div class="dhxsidebar_item_text">#text#</div></i>';
+    this.ui = this.root.attachSidebar({
+      template: 'faTiles'
+    });
+    this.ui.attachEvent('onSelect', (id) => this.callEvent('SideBar', [id]));
+    this._load();
 
-	_load(){
-		const struct = {
-			items:[
-				{
-					id:"projects",
-					text: "Projects",
-					icon: "projects.png",
-					selected: true
-				},
-				{
-					id:"about",
-					text: "About",
-					icon: "settings.png"
-				}
-			]
-		};
+    this.addService('SidebarService', {
+      select: (id) => {
+        this._select(id);
+      }
+    });
+  }
 
-		this.ui.loadStruct(struct);
-	}
-	
+  _select(id) {
+    this.ui.items(id).setActive();
+  }
+
+  _load() {
+    const struct = {
+      items: [
+        {
+          id: 'contacts',
+          text: 'Contacts',
+          icon: 'fa fa-user',
+          selected: true
+        },
+        {
+          id: 'projects',
+          text: 'Projects',
+          icon: 'fa fa-briefcase'
+        },
+        {
+          id: 'events',
+          text: 'Events',
+          icon: 'fa fa-calendar-check'
+        },
+        {
+          id: 'settings',
+          text: 'Settings',
+          icon: 'fa fa-cog'
+        },
+        {
+          id: 'about',
+          text: 'About',
+          icon: 'fa fa-info'
+        }
+      ]
+    };
+
+    this.ui.loadStruct(struct);
+  }
+
 }
